pyproject.toml
[tool.poetry]
name = "scuzzie"
version = "0.1.0"
description = "Simple static webcomic site generator."
authors = ["backwardspy <backwardspy@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
scuzzie = "scuzzie.cli:scuzzie"

[tool.poetry.dependencies]
python = "^3.9"
click = "^7.1.2"
toml = "^0.10.2"
Mako = "^1.1.4"
python-slugify = "^4.0.1"
marshmallow = "^3.10.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
flake8 = "^3.8.4"
mypy = "^0.812"
black = "^20.8b1"
ptpython = "^3.0.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
